requirements.txt
asgiref==3.5.2
astroid==2.12.10
contourpy==1.0.5
coverage==6.5.0
cycler==0.11.0
dill==0.3.5.1
distlib==0.3.5
Django==4.1.2
django-admin-action-buttons==0.1.9
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.9.0
django-admin-searchable-dropdown==1.2
django-autocomplete-light==3.9.4
django-coverage==1.2.4
django-six==1.0.4
django-utils-six==2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.9.0
lazy-object-proxy==1.9.0
mock==4.0.3
numpy==1.24.1
openpyxl==3.1.2
pandas==2.0.1
platformdirs==2.6.2
psycopg2==2.9.4
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.5
typing_extensions==4.3.0
tzdata==2023.3
virtualenv==20.16.3
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.0
